feat: Add file size and checksum to firmware uploads
This commit fixes a bug where the 'file_size' and 'checksum' fields were not being set when uploading a new firmware version. This caused a SQL error because the 'file_size' field does not have a default value.

The 'FirmwareController' has been updated to calculate the file size and checksum from the uploaded file and include them in the 'create' call.